requirements_test.txt
pytest>=8.0.0
pytest-homeassistant-custom-component>=0.12.0
